chore(pipeline): Migrates release script to ESM
